refactor(mcp-resources): extract kb:// URI surface + handlers (#157 step 2)
Step 2 of #157 (after #163) — extract the MCP resources surface from
src/KnowledgeBaseServer.ts so the kb:// scheme + list/read handlers
have their own seam.

src/mcp-resources.ts (new):
- mimeTypeForResource, buildResourceUri (renamed from local
  resourceUri), parseKnowledgeBaseResourceUri — pure URI helpers.
- listResources(), readResource(uri) — pure handler bodies; read
  KNOWLEDGE_BASES_ROOT_DIR from config like kb-stats.ts does.
- registerResources(mcp) — wires resources/list, resources/read, and
  the empty resources/templates/list onto an McpServer. Called once
  from KnowledgeBaseServer.buildMcpServer.

src/KnowledgeBaseServer.ts (-144 lines):
- handleListResources / handleReadResource remain as 1-line delegates
  so the existing private-method test surface keeps working without
  retooling KnowledgeBaseServer.test.ts.
- The local mimeTypeForResource / resourceUri / parseKnowledgeBase-
  ResourceUri / private registerResources are gone.

Cumulative across #157 step 1 (#163) + step 2: 1052 → 795 lines.

Adds src/mcp-resources.test.ts (19 cases) covering the pure URI
helpers in isolation: extension table, percent-encoding round-trip,
scheme/KB-name/path-segment validation, traversal rejection both
pre- and post-decode. The end-to-end resources path is still covered
by the 7 existing handleListResources/handleReadResource cases in
KnowledgeBaseServer.test.ts.

Refs #157 (step 3+ — manager-registry, warmup-fanout, sanitize-
MetadataForWire cleanup — to follow as separate PRs).
